Add features
HS is short for Harmonized System. The full name is "Harmonized Commodity Description and Coding System". As the name implies, each commodity is given a unique code, depending on its unique description. Therefore I can easily map commodity descriptions out of codes. For this, I use an excel file that already does this task. This file is publicly free and available on OEC website (short for Observatory of Economic Complexity).
First i read the excel files and convert them into proper dictionaries. Then for each hierarchy level of commodity codes, I map relative descriptions using the dictionaries as reference, and fill any remaining missing value with valid information manually.

Creating a column for regions adds some granularity to the countries with which products are traded. This is useful when it comes to creating a treemap for these countries, because it will allow us to cluster them into their proper regions, adding some beauty to the treemap and helping us analyze it more easily.